[ fix ] Profile 객체 -> String으로 바로 찾기
